clearing main.go move every function to helpers
